test: split TableParameters tests into 3 separate files
